requirements.txt
# Core dependencies
crewai>=0.11.0,<0.127.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.5.0

# Authentication dependencies
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Optional dependencies
requests>=2.31.0
tqdm>=4.67.0